Add files via upload
This is a Myntra Clone built using HTML and CSS. The project aims to replicate the user interface (UI) of the popular e-commerce website, Myntra. It includes core web pages such as the homepage, product listings, and a mock shopping cart. The project is fully responsive, providing a seamless experience across various screen sizes and devices.